add necessory to run seeds.js
