+++ package.json
@@ -5,7 +5,9 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "watch": "nodemon server.js",
+    "seed": "node seeds/seeds.js"
   },
   "author": "",
   "license": "ISC",
@@ -21,5 +23,8 @@
     "mysql2": "^2.3.3",
     "path": "^0.12.7",
     "sequelize": "^6.17.0"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.15"
   }
 }
